tests for lex and lite parser
